+++ R/add_race_sex.R
@@ -4,14 +4,14 @@ NULL
 
 #' Add columns for race/ethnicity and sex
 #'
-#' Add variables \code{race} and \code{sex} from \code{record} to a data
+#' Add variables \code{race} and \code{sex} from \code{dbase} to a data
 #' frame, using \code{id} is the join-by variable.
 #'
 #' Existing variables with the same names, if any, are overwritten.
 #'
 #' @param dframe data frame
 #' @param ... not used, forces later arguments to be used by name
-#' @param record data frame of student attributes, default midfieldstudents
+#' @param dbase data frame of student attributes, default midfieldstudents
 #' @return Data frame with the following properties:
 #' \itemize{
 #'     \item Rows are not modified
@@ -25,10 +25,10 @@ NULL
 #' # TBD
 add_race_sex <- function(dframe,
                          ...,
-                         record = NULL) {
+                         dbase = NULL) {
 
     # default arguments if NULL
-    record <- record %||% midfielddata::midfieldstudents
+    dbase <- dbase %||% midfielddata::midfieldstudents
 
     # bind names due to nonstandard evaluation notes in R CMD check
     # var <- NULL
@@ -36,19 +36,19 @@ add_race_sex <- function(dframe,
     # check arguments
     assert_explicit(dframe)
     assert_class(dframe, "data.frame")
-    assert_class(record, "data.frame")
+    assert_class(dbase, "data.frame")
 
     # existence of required columns
     assert_required_column(dframe, "id")
-    assert_required_column(record, "id")
-    assert_required_column(record, "race")
-    assert_required_column(record, "sex")
+    assert_required_column(dbase, "id")
+    assert_required_column(dbase, "race")
+    assert_required_column(dbase, "sex")
 
     # class of required columns
     assert_class(dframe[, id], "character")
-    assert_class(record[, id], "character")
-    assert_class(record[, race], "character")
-    assert_class(record[, sex], "character")
+    assert_class(dbase[, id], "character")
+    assert_class(dbase[, race], "character")
+    assert_class(dbase[, sex], "character")
 
     # preserve original data.frame, data.table, or tibble class
     df_class <- get_dframe_class(dframe)
@@ -61,10 +61,10 @@ add_race_sex <- function(dframe,
     key_names <- names_dframe[!names_dframe %chin% added_cols]
     dframe <- dframe[, key_names, with = FALSE]
 
-    # work on the record data
+    # work on the dbase data
     cols_we_want <- c("id", added_cols)
-    rows_we_want <- record$id %chin% dframe$id
-    DT <- record[rows_we_want, cols_we_want, with = FALSE]
+    rows_we_want <- dbase$id %chin% dframe$id
+    DT <- dbase[rows_we_want, cols_we_want, with = FALSE]
 
     # left-outer join, keep all rows of dframe
     dframe <-  merge(dframe, DT, by = "id", all.x = TRUE)