+++ R/filter_degree_seeking.R
@@ -4,36 +4,35 @@ NULL
 
 #' Subset a data frame for matriculants
 #'
-#' Subset a data frame by student ID, retaining students who are present in
-#' the matriculation record.
+#' Subset a data frame by student ID, retaining students who are degree-seeking.
 #'
-#' An inner join between \code{dframe} and the ID column of \code{record}
+#' An inner join between \code{dframe} and the ID column of \code{dbase}
 #' retains all rows of \code{dframe} with matching IDs.
 #'
 #' @param dframe data frame with an ID column
 #' @param ... not used, forces later arguments to be used by name
-#' @param record data frame of student attributes, default midfieldstudents
+#' @param dbase data frame of student attributes, default midfieldstudents
 #' @return Data frame with the following properties:
 #' \itemize{
-#'     \item Rows with matching IDs in \code{record}
+#'     \item Rows with matching IDs in \code{dbase}
 #'     \item Columns are not modified
 #'     \item Data frame attributes \code{tbl} or \code{data.table}
 #'           are preserved
 #'     \item Grouping structures are not preserved
 #' }
 #' @seealso \code{add_race_sex()}, \code{add_term_timely()},
-#'          \code{add_fair_record()}, \code{add_completion_timely()}
+#'          \code{add_data_sufficient()}, \code{add_completion_timely()}
 #' @export
 #' @examples
 #' # TBD
-subset_matriculant <- function(dframe, ..., record = NULL) {
+filter_degree_seeking <- function(dframe, ..., dbase = NULL) {
 
     wrapr::stop_if_dot_args(
         substitute(list(...)), "Arguments after ... must be named,"
     )
 
     # default arguments if NULL
-    record <- record %||% midfielddata::midfieldstudents
+    dbase <- dbase %||% midfielddata::midfieldstudents
 
     # bind names due to nonstandard evaluation notes in R CMD check
     # var <- NULL
@@ -41,15 +40,15 @@ subset_matriculant <- function(dframe, ..., record = NULL) {
     # check arguments
     assert_explicit(dframe)
     assert_class(dframe, "data.frame")
-    assert_class(record, "data.frame")
+    assert_class(dbase, "data.frame")
 
     # existence of required columns
     assert_required_column(dframe, "id")
-    assert_required_column(record, "id")
+    assert_required_column(dbase, "id")
 
     # class of required columns
     assert_class(dframe$id, "character")
-    assert_class(record$id, "character")
+    assert_class(dbase$id, "character")
 
     # preserve original data.frame, data.table, or tibble class
     df_class <- get_dframe_class(dframe)
@@ -59,9 +58,9 @@ subset_matriculant <- function(dframe, ..., record = NULL) {
     setDT(dframe)
     key_names <- names(dframe)
 
-    # work with record
-    # IDs in the record of matriculants
-    DT <- record[, .(id)]
+    # work with dbase
+    # IDs in the dbase of matriculants
+    DT <- dbase[, .(id)]
 
     # inner join, omit students not in matriculation data
     setkeyv(DT, "id")