+++ R/utils.R
@@ -7,13 +7,13 @@ NULL
 
 #' Add term range by institution
 #'
-#' Determine the latest academic term by institution in \code{record}.
+#' Determine the latest academic term by institution in \code{dbase}.
 #' Left-join by institution to \code{dframe} in a new column \code{inst_limit}.
 #'
 #' @param dframe data frame that received added column
-#' @param record data frame of term attributes
+#' @param dbase data frame of term attributes
 #' @noRd
-add_inst_limit <- function(dframe, record) {
+add_inst_limit <- function(dframe, dbase) {
 
   # bind names due to nonstandard evaluation notes in R CMD check
   # key_names <- NULL
@@ -28,7 +28,7 @@ add_inst_limit <- function(dframe, record) {
 
   # get max term by institution
   cols_we_want <- c("institution", "term")
-  DT <- record[, cols_we_want, with = FALSE]
+  DT <- dbase[, cols_we_want, with = FALSE]
   DT <- DT[, list(inst_limit = max(term)), by = "institution"]
 
   # left-outer join, keep all rows of dframe